requirements.txt
dj-database-url==0.4.1
Django==1.11.1
gunicorn==19.6.0
psycopg2==2.6.2
whitenoise==3.2
sendgrid==3.6.5
sendgrid-django==4.0.4
django-crispy-forms==1.6.1
django-datetime-widget==0.9.3
django-flat-responsive==2.0
django-ckeditor==5.3.0
Pillow==4.3.0
